# apps/docs/public/platform-api.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VhyxVoid platform API",
    "version": "1.0.0",
    "description": "The dashboard's API, for scripts, CI and editors. Authenticate with an API key: `Authorization: Bearer <keyId>.<secret>`. A key acts in its own workspace, as the member who created it, and each route needs one scope. Each key may make a number of requests per minute (plan limit); every response carries X-RateLimit headers. Generated from the routes the API serves."
  },
  "servers": [
    {
      "url": "https://api.vhyxvoid.com"
    }
  ],
...
  "paths": {
...
    "/api/v1/team/{accountId}/issues": {
      "get": {
        "operationId": "getTeamByAccountIdIssues",
        "tags": [
          "Team space"
        ],
        "summary": "List issues",
        "description": "q uses the filter syntax of the dashboard, for example: status:todo assignee:me label:bug.\n\nNeeds the `team:read` scope.",
        "x-required-scope": "team:read",
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "default": "",
              "type": "string",
              "maxLength": 500
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "default": "updated",
              "type": "string",
              "enum": [
                "rank",
                "updated",
                "created",
                "priority",
                "due",
                "number"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "headers": {
              "X-RateLimit-Limit": {
                "$ref": "#/components/headers/X-RateLimit-Limit"
              },
              "X-RateLimit-Remaining": {
                "$ref": "#/components/headers/X-RateLimit-Remaining"
              },
              "X-RateLimit-Reset": {
                "$ref": "#/components/headers/X-RateLimit-Reset"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/E400"
          },
          "401": {
            "$ref": "#/components/responses/E401"
          },
          "402": {
            "$ref": "#/components/responses/E402"
          },
          "403": {
            "$ref": "#/components/responses/E403"
          },
          "404": {
            "$ref": "#/components/responses/E404"
          },
          "429": {
            "$ref": "#/components/responses/E429"
          }
        }
      },
      "post": {
        "operationId": "postTeamByAccountIdIssues",
        "tags": [
          "Team space"
        ],
        "summary": "Open an issue",
        "description": "Opened by the key's creator. links attach dashboard objects (mock endpoints, requests, runs, docs).\n\nNeeds the `team:write` scope.",
        "x-required-scope": "team:write",
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "body": {
                    "default": "",
                    "type": "string",
                    "maxLength": 50000
                  },
                  "status": {
                    "default": "TODO",
                    "type": "string",
                    "enum": [
                      "BACKLOG",
                      "TODO",
                      "IN_PROGRESS",
                      "IN_REVIEW",
                      "DONE",
                      "CANCELLED"
                    ]
                  },
                  "priority": {
                    "default": "NONE",
                    "type": "string",
                    "enum": [
                      "NONE",
                      "LOW",
                      "MEDIUM",
                      "HIGH",
                      "URGENT"
                    ]
                  },
                  "assigneeId": {
                    "default": null,
                    "anyOf": [
                      {
                        "type": "string",
                        "format": "uuid"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "labels": {
                    "maxItems": 10,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "maxLength": 30
                    }
                  },
                  "dueDate": {
                    "default": null,
                    "anyOf": [
                      {
                        "type": "string",
                        "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "links": {
                    "default": [],
                    "maxItems": 20,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "maxLength": 2000
                    }
                  }
                },
                "required": [
                  "title"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success.",
            "headers": {
              "X-RateLimit-Limit": {
                "$ref": "#/components/headers/X-RateLimit-Limit"
              },
              "X-RateLimit-Remaining": {
                "$ref": "#/components/headers/X-RateLimit-Remaining"
              },
              "X-RateLimit-Reset": {
                "$ref": "#/components/headers/X-RateLimit-Reset"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/E400"
          },
          "401": {
            "$ref": "#/components/responses/E401"
          },
          "402": {
            "$ref": "#/components/responses/E402"
          },
          "403": {
            "$ref": "#/components/responses/E403"
          },
          "404": {
            "$ref": "#/components/responses/E404"
          },
          "429": {
            "$ref": "#/components/responses/E429"
          }
        }
      }
    },
...
    "/api/v1/team/{accountId}/issues/{n}/move": {
      "post": {
        "operationId": "postTeamByAccountIdIssuesByNMove",
        "tags": [
          "Team space"
        ],
        "summary": "Move an issue on the board",
        "description": "To a status column, between two issues (their numbers).\n\nNeeds the `team:write` scope.",
        "x-required-scope": "team:write",
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "n",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "BACKLOG",
                      "TODO",
                      "IN_PROGRESS",
                      "IN_REVIEW",
                      "DONE",
                      "CANCELLED"
                    ]
                  },
                  "before": {
                    "anyOf": [
                      {
                        "type": "integer",
                        "minimum": 1
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "after": {
                    "anyOf": [
                      {
                        "type": "integer",
                        "minimum": 1
                      },
                      {
                        "type": "null"
                      }
                    ]
                  }
                },
                "required": [
                  "status"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success.",
            "headers": {
              "X-RateLimit-Limit": {
                "$ref": "#/components/headers/X-RateLimit-Limit"
              },
              "X-RateLimit-Remaining": {
                "$ref": "#/components/headers/X-RateLimit-Remaining"
              },
              "X-RateLimit-Reset": {
                "$ref": "#/components/headers/X-RateLimit-Reset"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/E400"
          },
          "401": {
            "$ref": "#/components/responses/E401"
          },
          "402": {
            "$ref": "#/components/responses/E402"
          },
          "403": {
            "$ref": "#/components/responses/E403"
          },
          "404": {
            "$ref": "#/components/responses/E404"
          },
          "429": {
            "$ref": "#/components/responses/E429"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "keyId.secret",
        "description": "An API key from the dashboard: the key ID, a dot, and the secret."
      }
    },
    "headers": {
      "X-RateLimit-Limit": {
        "description": "The key's requests per minute.",
        "schema": {
          "type": "integer"
        }
      },
      "X-RateLimit-Remaining": {
        "description": "Requests left this minute.",
        "schema": {
          "type": "integer"
        }
      },
      "X-RateLimit-Reset": {
        "description": "Seconds until the minute ends.",
        "schema": {
          "type": "integer"
        }
      },
      "Retry-After": {
        "description": "Seconds to wait before trying again.",
        "schema": {
          "type": "integer"
        }
      }
    },
    "schemas": {
      "Success": {
        "type": "object",
        "required": [
          "success",
          "message",
          "data"
        ],
        "properties": {
          "success": {
            "const": true
          },
          "message": {
            "type": "string"
          },
          "data": {
            "description": "The result; its shape is in the route's description and the docs."
          }
        }
      },
...
    },
    "responses": {
      "E400": {
        "description": "The request is not valid; the message says why.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "E401": {
        "description": "The key is missing, wrong, revoked or expired, or this route needs a dashboard sign-in.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "E402": {
        "description": "A plan limit was reached.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "E403": {
        "description": "The key lacks the route's scope, belongs to another workspace, or its creator is no longer a member.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "E404": {
        "description": "Not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "E429": {
        "description": "The key's requests-per-minute limit (or the per-address limit) was reached. Retry after Retry-After.",
        "headers": {
          "Retry-After": {
            "$ref": "#/components/headers/Retry-After"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  }
}